requirements.txt
praw
nltk
textblob
fpdf2